feat: 在npu推理前后暂停vpss流水线，验证内存带宽冲突问题

+++ src/rknn/yolov5_model.cpp
@@ -189,8 +189,8 @@ int YoloV5Model::Init(const ModelConfig& config) {
                  input_attrs_[i].dims[0], input_attrs_[i].dims[1], 
                  input_attrs_[i].dims[2], input_attrs_[i].dims[3],
                  input_attrs_[i].n_elems, input_attrs_[i].size, input_attrs_[i].size_with_stride,
-                 input_attrs_[i].fmt, input_attrs_[i].type,
-                 input_attrs_[i].qnt_type, input_attrs_[i].zp, input_attrs_[i].scale);
+                 (int)input_attrs_[i].fmt, (int)input_attrs_[i].type,
+                 (int)input_attrs_[i].qnt_type, input_attrs_[i].zp, input_attrs_[i].scale);
     }
     
     // 4. 查询输出张量属性
@@ -211,8 +211,8 @@ int YoloV5Model::Init(const ModelConfig& config) {
                  output_attrs_[i].dims[0], output_attrs_[i].dims[1], 
                  output_attrs_[i].dims[2], output_attrs_[i].dims[3],
                  output_attrs_[i].n_elems, output_attrs_[i].size, output_attrs_[i].size_with_stride,
-                 output_attrs_[i].fmt, output_attrs_[i].type,
-                 output_attrs_[i].qnt_type, output_attrs_[i].zp, output_attrs_[i].scale);
+                 (int)output_attrs_[i].fmt, (int)output_attrs_[i].type,
+                 (int)output_attrs_[i].qnt_type, output_attrs_[i].zp, output_attrs_[i].scale);
     }
     
     // 5. 设置输入类型为 UINT8，使用 NHWC 格式（RV1106 零拷贝模式要求）
